requirements.txt
# Core framework dependency
mcp-agent @ file://../../../  # Link to the local mcp-agent project root

# Additional dependencies specific to this example
anthropic
openai

# Streamlit and web app dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0

# Database (SQLite3 is included in Python standard library)

# Optional: For enhanced authentication
streamlit-authenticator>=0.2.3

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
ruff>=0.1.0
mypy>=1.5.0